Motion: don't prebuffer when we're recording
A lot of content was being duplicated because we kept the prebuffer
after a motion event. If another motion event occurred before the
prebuffer had cycled through, we'd duplicate the last bit of a prior
motion event.

We fix this by ending motion events on an IDR so we can seed the
prebuffer, and by clearing the prebuffer when a motion event begins.

Signed-off-by: Nick High <nick@i386.io>